% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inv_sqrt}
\alias{inv_sqrt}
\title{Inverse square root of a symmetric matrix}
\usage{
inv_sqrt(L_sym)
}
\arguments{
\item{L_sym}{A named, symmetric numeric matrix.}
}
\value{
A symmetric numeric matrix of the same dimensions as \code{L_sym},
representing \eqn{L_{\text{sym}}^{-1/2}}.
}
\description{
Computes the inverse square root of a symmetric positive (semi-)definite
matrix using its eigen-decomposition.
}
\keyword{internal}
